+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - fixed logging
 - fix no branch crash when no branch is set up 
 - added fetch buttons
+- double ESC will exit gitonic
 -